# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Gloo",
        "title": "Gloo API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@gloo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "gloo-api-production.up.railway.app",
    "basePath": "/api",
    "paths": {
...
        "/follows/{userId}": {
            "post": {
                "description": "Create a new follow",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Follows"
                ],
                "summary": "Create a follow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Follow",
                        "name": "follow",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateFollowRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/follows/{userId}/{followId}": {
            "delete": {
                "description": "Delete a follow",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "follows"
                ],
                "summary": "Delete a follow",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Follow ID",
                        "name": "followId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
        "/ingredients/{recipeId}/{ingredientId}": {
            "put": {
                "description": "Update a specific ingredient by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ingredients"
                ],
                "summary": "Update ingredient",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "recipeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Ingredient ID",
                        "name": "ingredientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ingredient",
                        "name": "ingredient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateIngredientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated ingredient",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Ingredient not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new ingredient",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ingredients"
                ],
                "summary": "Create ingredient",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "recipeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Ingredient ID",
                        "name": "ingredientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Ingredient",
                        "name": "ingredient",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateIngredientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created ingredient",
                        "schema": {
                            "$ref": "#/definitions/models.CreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Ingredient not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a specific ingredient by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ingredients"
                ],
                "summary": "Delete ingredient",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Recipe ID",
                        "name": "recipeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Ingredient ID",
                        "name": "ingredientId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted ingredient",
                        "schema": {
                            "$ref": "#/definitions/models.DeleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorBadRequestResponse"
                        }
                    },
                    "404": {
                        "description": "Ingredient not found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorNotFoundResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorInternalServerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateFollowRequest": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.CreateIngredientRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "onion"
                },
                "name": {
                    "type": "string",
                    "example": "Chicken"
                },
                "quantity": {
                    "type": "string",
                    "example": "100"
                },
                "unit": {
                    "type": "string",
                    "example": "kg"
                }
            }
        },
...
        "models.CreateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Created successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 201
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.DeleteResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Deleted successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.ErrorBadRequestResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Bad Request"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 400
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.ErrorInternalServerErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 500
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.ErrorNotFoundResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Not Found"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "status_code": {
                    "type": "integer",
                    "example": 404
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "models.UpdateIngredientRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "onion"
                },
                "name": {
                    "type": "string",
                    "example": "Chicken"
                },
                "quantity": {
                    "type": "string",
                    "example": "100"
                },
                "unit": {
                    "type": "string",
                    "example": "kg"
                }
            }
        },
...
        "models.UpdateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Updated successfully"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}